Move several files from `payment/` to `order/`
