+++ src/features/order/components/OrderInfo/index.js
@@ -1,0 +1,1 @@
+export { default as OrderInfo } from "./OrderInfo";
