+++ src/features/payment/components/CheckoutSection/index.js
@@ -1,0 +1,1 @@
+export { default as PaymentCheckoutSection } from "./PaymentCheckoutSection";
