+++ src/features/payment/components/Container/index.js
@@ -1,0 +1,1 @@
+export * from "./PaymentGrids";
