+++ src/features/payment/components/Info/ProductBill.jsx
@@ -10,7 +10,7 @@ import {
 } from "@mui/material";
 import { SpecsTableRow, SpecsTableCell } from "components/Table";
 import CartItem from "features/payment/components/Info/CartItem";
-import { orderType } from "features/payment/types";
+import { orderType } from "features/order/types";
 import { useCheckoutPrices } from "hooks";
 import formatNumberToVND from "utils/currency-formatter";
