+++ src/features/payment/components/Info/index.js
@@ -1,4 +1,3 @@
 export { default as CartItemList } from "./CartItemList";
 export { default as DeliveryInfo } from "./DeliveryInfo";
-export { default as OrderInfo } from "./OrderInfo";
 export { default as SupportPaymentTypes } from "./SupportPaymentTypes";
