+++ src/features/payment/types/index.js
@@ -1,2 +1,1 @@
 export * from "./cart-item";
-export * from "./order";
